Add files via upload
This is the defensive module of the project, which is developed using .NET framework. A windows application is developed consisting of numerous modules, such as port checker, system scanner, process analyzer etc. The function of this application is to identify the existing undetectable file in the system and remove it.